+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/BaseSentryNonAndroidPluginTest.kt
@@ -21,6 +21,7 @@ abstract class BaseSentryNonAndroidPluginTest(
     private lateinit var rootBuildFile: File
     protected lateinit var appBuildFile: File
     protected lateinit var moduleBuildFile: File
+    protected lateinit var sentryPropertiesFile: File
     protected lateinit var runner: GradleRunner
 
     protected open val additionalRootProjectConfig: String = ""
@@ -36,6 +37,7 @@ abstract class BaseSentryNonAndroidPluginTest(
 
         appBuildFile = File(testProjectDir.root, "app/build.gradle")
         moduleBuildFile = File(testProjectDir.root, "module/build.gradle")
+        sentryPropertiesFile = File(testProjectDir.root, "sentry.properties")
         rootBuildFile = testProjectDir.writeFile("build.gradle") {
             // language=Groovy
             """